HW3-2 | Float to Double conversion of array

+++ HW3/HW3-220223055_2.c
@@ -23,46 +23,49 @@
 // a0 + a1x + a2x2 + … + aNxN
 //
 
-float PolinomHesapla(float *p, int N, float x) 
+#include <stdio.h>
+#include <math.h>
+
+double PolinomHesapla(double *p, int N, double x) 
 {
     int i;
-    float sonuc = 0;
+    double sonuc = 0;
 
-    	for (i=0; i<=N; i++)
-		{
-     	   sonuc += *(p + i) * pow(x, i);
-   		}
-    
+    for (i = 0; i <= N; i++)
+    {
+        sonuc += *(p + i) * pow(x, (double)i);
+    }
+
     return sonuc;
 }
 
 int main() 
 {
     int N, i;
-    float x;
+    double x;
     
     printf("---------------------------------------\n");
     printf("Polinomun derecesini girin: ");
     scanf("%d", &N);
     printf("----------------------------------------\n");
-  	printf("----\n"); 	    
+    printf("----\n"); 	    
 
-    float p[N + 1];
+    double p[N + 1];
 
-    for (i=0; i<=N; i++) 
-	{
-        printf("%d. Degiskenin katsayisini girin: ", i+1);
-        scanf("%f", &p[i]);
-  	    printf("----\n");            
+    for (i = 0; i <= N; i++) 
+    {
+        printf("%d. Degiskenin katsayisini girin: ", i + 1);
+        scanf("%lf", &p[i]);
+        printf("----\n");            
     }
 
     printf("\nPolinomun degerinin hesaplanacagi noktayi girin: ");
-    scanf("%f", &x);
+    scanf("%lf", &x);
 
-    float sonuc = PolinomHesapla(p, N, x);
-  	printf("\n-----------------------\n");
-    printf("\nSonuc: %.2f\n", sonuc);
-  	printf("\n-----------------------\n");    
+    double sonuc = PolinomHesapla(p, N, x);
+    printf("\n-----------------------\n");
+    printf("\nSonuc: %.2lf\n", sonuc);
+    printf("\n-----------------------\n");    
 
     return 0;
 }
